+++ test/unit/mysql_rep_admin/main.py
@@ -9,7 +9,6 @@
         test/unit/mysql_rep_admin/main.py
 
     Arguments:
-        None
 
 """
 
@@ -30,9 +29,37 @@ import mock
 # Local
 sys.path.append(os.getcwd())
 import mysql_rep_admin
+import lib.gen_libs as gen_libs
 import version
 
 __version__ = version.__version__
+
+
+class ProgramLock(object):
+
+    """Class:  ProgramLock
+
+    Description:  Class stub holder for gen_class.ProgramLock class.
+
+    Methods:
+        __init__ -> Class initialization.
+
+    """
+
+    def __init__(self, cmdline, flavor):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+            (input) cmdline -> Argv command line.
+            (input) flavor -> Lock flavor ID.
+
+        """
+
+        self.cmdline = cmdline
+        self.flavor = flavor
 
 
 class UnitTest(unittest.TestCase):
@@ -55,6 +82,10 @@ class UnitTest(unittest.TestCase):
         test_arg_dir_false -> Test arg_dir_chk_crt if returns false.
         test_arg_file_true -> Test arg_file_chk if returns true.
         test_arg_file_false -> Test arg_file_chk if returns false.
+        test_run_program -> Test run_program function.
+        test_programlock_true -> Test with ProgramLock returns True.
+        test_programlock_false -> Test with ProgramLock returns False.
+        test_programlock_id -> Test ProgramLock with flavor ID.
 
     """
 
@@ -68,7 +99,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.args_array = {"-c": "CfgFile", "-d": "CfgDir", "-C": True}
+        self.args_array = {"-c": "CfgFile", "-d": "CfgDir", "-C": True,
+                           "-s": "slave.txt"}
+        self.args_array2 = {"-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor"}
+        self.proglock = ProgramLock(["cmdline"], "FlavorID")
 
     @mock.patch("mysql_rep_admin.gen_libs.help_func")
     @mock.patch("mysql_rep_admin.arg_parser.arg_parse2")
@@ -289,10 +323,11 @@ class UnitTest(unittest.TestCase):
 
         self.assertFalse(mysql_rep_admin.main())
 
-    @mock.patch("mysql_rep_admin.run_program")
+    @mock.patch("mysql_rep_admin.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_admin.gen_class.ProgramLock")
     @mock.patch("mysql_rep_admin.gen_libs.help_func")
     @mock.patch("mysql_rep_admin.arg_parser")
-    def test_arg_file_false(self, mock_arg, mock_help, mock_run):
+    def test_arg_file_false(self, mock_arg, mock_help, mock_lock):
 
         """Function:  test_arg_file_false
 
@@ -309,7 +344,109 @@ class UnitTest(unittest.TestCase):
         mock_arg.arg_file_chk.return_value = True
         mock_arg.arg_dir_chk_crt.return_value = False
         mock_arg.arg_file_chk.return_value = False
-        mock_run.return_value = True
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_rep_admin.main())
+
+    @mock.patch("mysql_rep_admin.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_admin.gen_class.ProgramLock")
+    @mock.patch("mysql_rep_admin.gen_libs.help_func")
+    @mock.patch("mysql_rep_admin.arg_parser")
+    def test_run_program(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_run_program
+
+        Description:  Test run_program function.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_help.return_value = False
+        mock_arg.arg_req_or_lst.return_value = True
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_file_chk.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_rep_admin.main())
+
+    @mock.patch("mysql_rep_admin.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_admin.gen_class.ProgramLock")
+    @mock.patch("mysql_rep_admin.gen_libs.help_func")
+    @mock.patch("mysql_rep_admin.arg_parser")
+    def test_programlock_true(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_programlock_true
+
+        Description:  Test with ProgramLock returns True.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_help.return_value = False
+        mock_arg.arg_req_or_lst.return_value = True
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_file_chk.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_rep_admin.main())
+
+    @mock.patch("mysql_rep_admin.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_admin.gen_class.ProgramLock")
+    @mock.patch("mysql_rep_admin.gen_libs.help_func")
+    @mock.patch("mysql_rep_admin.arg_parser")
+    def test_programlock_false(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_programlock_false
+
+        Description:  Test with ProgramLock returns False.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_help.return_value = False
+        mock_arg.arg_req_or_lst.return_value = True
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_file_chk.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.side_effect = \
+            mysql_rep_admin.gen_class.SingleInstanceException
+
+        with gen_libs.no_std_out():
+            self.assertFalse(mysql_rep_admin.main())
+
+    @mock.patch("mysql_rep_admin.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_admin.gen_class.ProgramLock")
+    @mock.patch("mysql_rep_admin.gen_libs.help_func")
+    @mock.patch("mysql_rep_admin.arg_parser")
+    def test_programlock_id(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_programlock_id
+
+        Description:  Test ProgramLock with flavor ID.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array2
+        mock_help.return_value = False
+        mock_arg.arg_req_or_lst.return_value = True
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_file_chk.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
 
         self.assertFalse(mysql_rep_admin.main())