+++ test/unit/mysql_rep_admin/process_json.py
@@ -58,7 +58,8 @@ class Mail(object):
 
         """
 
-        pass
+        self.lag_time = lag_time
+        self.data = None
 
     def add_2_msg(self, data):
 
@@ -69,6 +70,8 @@ class Mail(object):
         Arguments:
 
         """
+
+        self.data = data
 
         return True